+++ mobile/task/app/test/data/models/currency_model_test.dart
@@ -1,7 +1,7 @@
 import 'package:app/core/utils.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../mock_data.dart';
+import 'package:app/core/mock_data.dart';
 import '../model_factory.dart';
 
 void main() {
