+++ mobile/task/app/test/data/service/currencies_service_test.dart
@@ -6,7 +6,7 @@ import 'package:app/data/service/currencies_service.dart';
 import 'package:app/data/models/currency_model.dart';
 import 'package:app/data/models/detailed_currency_model.dart';
 
-import '../mock_data.dart';
+import 'package:app/core/mock_data.dart';
 
 class MockGraphQLClient extends Mock implements GraphQLClient {}
