+++ mobile/task/app/test/domain/entity_factory.dart
@@ -2,7 +2,7 @@ import 'package:app/core/utils.dart';
 import 'package:app/domain/entities/currency.dart';
 import 'package:app/domain/entities/detailed_currency.dart';
 
-import '../data/mock_data.dart';
+import 'package:app/core/mock_data.dart';
 
 Currency createCurrency({
   String code = code,
