+++ mobile/task/app/test/presentation/screens/currency_detail_screen_test.dart
@@ -5,7 +5,7 @@ import 'package:app/presentation/widgets/rate_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../data/mock_data.dart';
+import 'package:app/core/mock_data.dart';
 import '../../domain/entity_factory.dart';
 import '../widget_test_utils.dart';
